Move run dcos command.
